Refactor paser and command in test
