+++ src/test/java/seedu/address/logic/parser/person/RemarkPersonCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.person;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_REMARK;
@@ -10,11 +10,10 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.person.RemarkPersonCommand;
-import seedu.address.logic.parser.person.RemarkCommandParser;
 import seedu.address.model.person.Remark;
 
 public class RemarkPersonCommandParserTest {
-    private RemarkCommandParser parser = new RemarkCommandParser();
+    private RemarkPersonCommandParser parser = new RemarkPersonCommandParser();
     private final String nonEmptyRemark = "Some remark.";
 
     @Test
